+++ src/components/image-header-text-link/image-header-text-link-component.js
@@ -3,24 +3,19 @@ import {flexColumn, flexRow, left, right} from './image-header-text-link-compone
 
 export default ImageHeader;
 
-ImageHeader.defaultProps = {
-    images: []
-};
+function ImageHeader({image, position, header, sentence, linkDestination, linkText}) {
 
-function ImageHeader({images}) {
-
-    const side = images.position === true ? right : left;
+    const side = position === true ? right : left;
     return (
         <div className={`${flexRow} ${side}`}>
-            {images.map(image => (<section>
             <div className={flexColumn}>
-                <img src={image.image}/>
+                <img src={image}/>
             </div>
             <div className={flexColumn}>
-                <h1>{image.header}</h1>
-                <p>{image.sentence}</p>
-                <a href={image.linkDestination}>{image.linkText} &#8594;</a>
-            </div></section>))};
+                <h1>{header}</h1>
+                <p>{sentence}</p>
+                <a href={linkDestination}>{linkText} &#8594;</a>
+            </div>))};
         </div>
     );
