+++ src/pages/index.js
@@ -3,5 +3,7 @@ import {FullWidthImage} from '../components/images/images-component';
 import Layout from '../components/Layout';
 
 export default () =>  {
-    return <Layout><h1>Hello</h1></Layout>;
+    return <Layout>
+
+    </Layout>;
 }
